+++ src/main/java/play/pluv/member/domain/repository/MemberRepository.java
@@ -1,6 +1,7 @@
-package play.pluv.member.domain;
+package play.pluv.member.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import play.pluv.member.domain.Member;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
