+++ src/test/java/play/pluv/login/application/RegisterUpdaterTest.java
@@ -6,7 +6,7 @@ import static play.pluv.playlist.domain.MusicStreaming.SPOTIFY;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import play.pluv.member.domain.Member;
-import play.pluv.member.domain.MemberRepository;
+import play.pluv.member.domain.repository.MemberRepository;
 import play.pluv.oauth.domain.OAuthMemberInfo;
 import play.pluv.support.ApplicationTest;
